requirements.txt
discord.py==2.3.2
nextcord==2.6.0
nltk==3.8.1
Pillow==10.2.0
pygame==2.5.2
selenium==4.16.0
wikipedia==1.4.0
Wikipedia_API==0.6.0
